requirements.txt
# Core
ccxt>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0

# AI/ML
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
joblib>=1.3.0

# Web Dashboard
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Scheduling
apscheduler>=3.10.0

# Notifications
requests>=2.31.0

# Security
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
